requirements.txt
Flask>=3.0.0
Flask-SocketIO>=5.3.6
simple-websocket>=1.0.0
pyserial>=3.5
numpy>=1.24.0
Pillow>=10.0.0
PyYAML>=6.0
# Optional — install only if using real cameras:
# opencv-python-headless>=4.8.0
# Optional — install only if using real YDLIDAR X2:
# ydlidar-sdk>=1.0.0
